PoseTransOpt/requirements.txt
human_body_prior==2.2.2.0
icecream==2.1.3
imageio==2.31.2
matplotlib==3.5.3
numpy==1.21.5
open3d==0.16.0
opencv_python==4.8.0.76
pyrender==0.1.45
scipy==1.7.3
torch==1.12.1
torchgeometry==0.1.2
tqdm==4.65.2
trimesh==3.23.5
vctoolkit==0.1.16.0
hydra-core==1.3.0
